Extract MemoApp class
